Question_4/requirements.txt
numpy==1.26.3
Flask==3.0.3
pytest==8.1.1
pillow==10.2.0